coerce apply tests to CRD v1. skip CRDs that are no longer valid
